chore: enhasment in clear yarn command

+++ _enabled-commands/clear-pods.sh
@@ -9,6 +9,7 @@ export LANG=en_US.UTF-8
 # Optional parameters:
 # @raycast.icon 🤖
 # @raycast.argument1 { "type": "text", "placeholder": "clear node modules?" }
+# @raycast.argument2 { "type": "text", "placeholder": "clear pods?" }
 # @raycast.packageName frontend-universe
 # @raycast.needsConfirmation true
 
@@ -21,8 +22,12 @@ if [ "$1" = "yes" ]; then
   ./clear-yarn.sh
 fi
 
-cd ~/WebstormProjects/frontend-universe/packages/apps/classroom/ios
-rm -rf Pods/
+cd ~/WebstormProjects/frontend-universe/packages/apps/classroom/ios || exit 
+
+if [ "$2" = "yes" ]; then
+  rm -rf Pods/
+fi
+
 
 pod install
